+++ spark/v3.1/spark/src/test/java/com/netease/arctic/spark/TestKeyedTableDDL.java
@@ -142,62 +142,4 @@ public class TestKeyedTableDDL extends SparkTestBase {
     sql("drop table {0}.{1}", database, table);
     assertTableNotExist(identifier);
   }
-
-  @Test
-  public void testCreateNewTableShouldHaveTimestampWithoutZone() {
-    withSQLConf(ImmutableMap.of(
-            USE_TIMESTAMP_WITHOUT_TIME_ZONE_IN_NEW_TABLES, "true"), () -> {
-      TableIdentifier identifier = TableIdentifier.of(catalogNameArctic, database, table);
-
-      sql("create table {0}.{1} ( \n" +
-              " id int , \n" +
-              " name string , \n " +
-              " ts timestamp , \n" +
-              " primary key (id) \n" +
-              ") using arctic \n" +
-              " partitioned by ( days(ts) ) \n" +
-              " tblproperties ( \n" +
-              " ''props.test1'' = ''val1'', \n" +
-              " ''props.test2'' = ''val2'' ) ", database, table);
-      Types.StructType expectedSchema = Types.StructType.of(
-              Types.NestedField.required(1, "id", Types.IntegerType.get()),
-              Types.NestedField.optional(2, "name", Types.StringType.get()),
-              Types.NestedField.optional(3, "ts", Types.TimestampType.withoutZone()));
-      Assert.assertEquals("Schema should match expected",
-              expectedSchema, loadTable(identifier).schema().asStruct());
-      List<Object[]> values = ImmutableList.of(
-              row(1L, toTimestamp("2021-01-01T00:00:00.0"), toTimestamp("2021-02-01T00:00:00.0")),
-              row(2L, toTimestamp("2021-01-01T00:00:00.0"), toTimestamp("2021-02-01T00:00:00.0")),
-              row(3L, toTimestamp("2021-01-01T00:00:00.0"), toTimestamp("2021-02-01T00:00:00.0"))
-      );
-      sql("INSERT INTO {0}.{1} VALUES {2}", database, table, rowToSqlValues(values));
-
-      rows = sql("SELECT * FROM {0}.{1} ORDER BY id", database, table);
-      Assert.assertEquals(3, rows.size());
-      sql("drop table {0}.{1}", database, table);
-    });
-  }
-
-  private String rowToSqlValues(List<Object[]> rows) {
-    List<String> rowValues = rows.stream().map(row -> {
-      List<String> columns = Arrays.stream(row).map(value -> {
-        if (value instanceof Long) {
-          return value.toString();
-        } else if (value instanceof Timestamp) {
-          return String.format("timestamp '%s'", value);
-        }
-        throw new RuntimeException("Type is not supported");
-      }).collect(Collectors.toList());
-      return "(" + Joiner.on(",").join(columns) + ")";
-    }).collect(Collectors.toList());
-    return Joiner.on(",").join(rowValues);
-  }
-
-  protected Object[] row(Object... values) {
-    return values;
-  }
-
-  private Timestamp toTimestamp(String value) {
-    return new Timestamp(DateTime.parse(value).getMillis());
-  }
 }
